Crate multiply_by_rational_with_rounding

Running

Running this fuzzer can be done with cargo hfuzz run multiply_by_rational_with_rounding. honggfuzz CLI options can be used by setting HFUZZ_RUN_ARGS, such as -n 4 to use 4 threads.

Debugging a panic

Once a panic is found, it can be debugged with cargo hfuzz run-debug multiply_by_rational_with_rounding hfuzz_workspace/multiply_by_rational_with_rounding/*.fuzz.

More information

More information about honggfuzz can be found here.

Structs

ArbitraryRounding 🔒

An [arbitrary::Arbitrary] [Rounding] mode.

Functions

check 🔒

main 🔒

Tries to demonstrate that multiply_by_rational_with_rounding is incorrect.

round 🔒

Round a Fraction according to the given mode.